refactor(server): hoist isAsyncGenerator + readableStreamToSse into sse.ts
