+++ packages/server/src/actions-handler.ts
@@ -1,38 +1,11 @@
 import type { MiddlewareHandler } from 'hono';
 import { ActionGuardError, type ActionGuardFn, type ActionGuardContext } from '@hono-preact/iso';
 import { runRequestScope } from '@hono-preact/iso/internal';
-import { sseFromGenerator, sseEncode, sseEncodeError } from './sse.js';
-
-function isAsyncGenerator(value: unknown): value is AsyncGenerator<unknown, unknown, unknown> {
-  return (
-    value != null &&
-    typeof value === 'object' &&
-    typeof (value as { [Symbol.asyncIterator]?: unknown })[Symbol.asyncIterator] === 'function' &&
-    typeof (value as { next?: unknown }).next === 'function'
-  );
-}
-
-function readableStreamToSse(stream: ReadableStream<unknown>): ReadableStream<Uint8Array> {
-  const reader = stream.getReader();
-  return new ReadableStream<Uint8Array>({
-    async start(controller) {
-      try {
-        while (true) {
-          const { done, value } = await reader.read();
-          if (done) break;
-          controller.enqueue(sseEncode({ data: JSON.stringify(value) }));
-        }
-      } catch (err) {
-        controller.enqueue(sseEncodeError(err));
-      } finally {
-        controller.close();
-      }
-    },
-    cancel() {
-      reader.cancel().catch(() => { /* swallow */ });
-    },
-  });
-}
+import {
+  sseFromGenerator,
+  isAsyncGenerator,
+  readableStreamToSse,
+} from './sse.js';
 
 type GlobModule = {
   __moduleKey?: unknown;
