+++ packages/server/src/loaders-handler.ts
@@ -1,6 +1,10 @@
 import type { MiddlewareHandler } from 'hono';
 import { runRequestScope } from '@hono-preact/iso/internal';
-import { sseFromGenerator, sseEncode, sseEncodeError } from './sse.js';
+import {
+  sseFromGenerator,
+  isAsyncGenerator,
+  readableStreamToSse,
+} from './sse.js';
 
 type GlobModule = {
   default?: unknown;
@@ -49,37 +53,6 @@ function validateLocation(loc: unknown): SerializedLocation | null {
     pathParams: o.pathParams as Record<string, string>,
     searchParams: o.searchParams as Record<string, string>,
   };
-}
-
-function isAsyncGenerator(value: unknown): value is AsyncGenerator<unknown, unknown, unknown> {
-  return (
-    value != null &&
-    typeof value === 'object' &&
-    typeof (value as { [Symbol.asyncIterator]?: unknown })[Symbol.asyncIterator] === 'function' &&
-    typeof (value as { next?: unknown }).next === 'function'
-  );
-}
-
-function readableStreamToSse(stream: ReadableStream<unknown>): ReadableStream<Uint8Array> {
-  const reader = stream.getReader();
-  return new ReadableStream<Uint8Array>({
-    async start(controller) {
-      try {
-        while (true) {
-          const { done, value } = await reader.read();
-          if (done) break;
-          controller.enqueue(sseEncode({ data: JSON.stringify(value) }));
-        }
-      } catch (err) {
-        controller.enqueue(sseEncodeError(err));
-      } finally {
-        controller.close();
-      }
-    },
-    cancel() {
-      reader.cancel().catch(() => { /* swallow */ });
-    },
-  });
 }
 
 export function loadersHandler(glob: LazyGlob | EagerGlob): MiddlewareHandler {
